+++ backend/integration_tests/src/p2p_swap_tests.rs
@@ -102,7 +102,7 @@ fn p2p_swap_in_direct_chat_succeeds() {
 
     if let ChatEvent::Message(m) = user1_event {
         if let MessageContent::P2PSwap(p) = m.content {
-            assert!(matches!(p.status, P2PSwapStatus::Accepted(c) if c.accepted_by == user2.user_id));
+            assert!(matches!(p.status, P2PSwapStatus::Completed(c) if c.accepted_by == user2.user_id));
         }
     }
 
@@ -114,7 +114,7 @@ fn p2p_swap_in_direct_chat_succeeds() {
 
     if let ChatEvent::Message(m) = user2_event {
         if let MessageContent::P2PSwap(p) = m.content {
-            assert!(matches!(p.status, P2PSwapStatus::Accepted(c) if c.accepted_by == user2.user_id));
+            assert!(matches!(p.status, P2PSwapStatus::Completed(c) if c.accepted_by == user2.user_id));
         }
     }
 }